Add method print all comments for book by id

+++ hw06-hibernate/src/main/java/me/decentos/model/Book.java
@@ -34,7 +34,7 @@ public class Book {
         this.genre = genre;
     }
 
-    public Book byId(int id) {
+    public Book byId(long id) {
         this.id = id;
         return this;
     }
